+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 * Fix `DeprecationWarning` on invalid escape sequences (thanks @DavidCain)
+* Add validation for arguments in Address.delete, Address.find, and Address.update
 
 ## 4.16.0
 * Add `LiabilityShift` class and `liability_shift` to RiskData